refactor: generateTabIds
