+++ packages/core/src/components/index.ts
@@ -118,7 +118,7 @@ export { CheckboxCard, type CheckboxCardProps } from "./control-card/checkboxCar
 export { RadioCard, type RadioCardProps } from "./control-card/radioCard";
 export { SwitchCard, type SwitchCardProps } from "./control-card/switchCard";
 export { Tab, type TabId, type TabProps } from "./tabs/tab";
-export { TabPanel as TabPanelWrapper } from "./tabs/tabPanelWrapper";
+export { TabPanel as TabPanelWrapper } from "./tabs/tabPanel";
 // eslint-disable-next-line deprecation/deprecation
 export { Tabs, type TabsProps, TabsExpander, Expander } from "./tabs/tabs";
 export { CompoundTag, type CompoundTagProps } from "./tag/compoundTag";
