+++ packages/core/src/components/tabs/tabs.tsx
@@ -20,7 +20,7 @@ import * as React from "react";
 import { AbstractPureComponent, Classes, DISPLAYNAME_PREFIX, type Props, Utils } from "../../common";
 
 import { Tab, type TabId, type TabProps } from "./tab";
-import { TabPanel } from "./tabPanelWrapper";
+import { TabPanel } from "./tabPanel";
 import { TabTitle } from "./tabTitle";
 
 /**
