+++ packages/core/test/tabs/tabsTests.tsx
@@ -21,7 +21,7 @@ import { spy } from "sinon";
 import { Classes } from "../../src/common";
 import { Tab } from "../../src/components/tabs/tab";
 import { Tabs, type TabsProps, type TabsState } from "../../src/components/tabs/tabs";
-import { generateTabPanelId, generateTabTitleId } from "../../src/components/tabs/tabTitle";
+import { generateTabIds } from "../../src/components/tabs/tabTitle";
 
 describe("<Tabs>", () => {
     const ID = "tabsTests";
@@ -130,13 +130,14 @@ describe("<Tabs>", () => {
     });
 
     it("passes correct tabTitleId and tabPanelId to panel renderer", () => {
+        const expectedIds = generateTabIds(ID, "first");
         mount(
             <Tabs id={ID}>
                 <Tab
                     id="first"
                     panel={({ tabTitleId, tabPanelId }) => {
-                        assert.equal(tabTitleId, generateTabTitleId(ID, "first"));
-                        assert.equal(tabPanelId, generateTabPanelId(ID, "first"));
+                        assert.equal(tabTitleId, expectedIds.tabTitleId);
+                        assert.equal(tabPanelId, expectedIds.tabPanelId);
                         return <Panel title="a" />;
                     }}
                 />
